+++ ModelExplorerPlugin/LevelPropertyViewBuilder.h
@@ -10,6 +10,7 @@
 
 #include "PropertyViewBuilder.h"
 
+
 class LevelPropertyViewBuilder : public PropertyViewBuilder
 {
 public:
