+++ ModelExplorerPlugin/ModelExplorerPlugin.vcxproj
@@ -169,7 +169,7 @@ copy /y "$(ProjectDir)Resources\ModelExplorerIcon.png" "$(OutDir)ModelExplorerIc
     <ClCompile Include="ModelExplorerWidget.cpp" />
     <ClCompile Include="ObjectSelectionHandler.cpp" />
     <ClCompile Include="PluginToolButtons.cpp" />
-    <ClCompile Include="PropertyManagers.cpp" />
+    <ClCompile Include="PropertyManager.cpp" />
     <ClCompile Include="PropertyView.cpp" />
     <ClCompile Include="PropertyViewBuilder.cpp" />
     <ClCompile Include="RebarUsagePropertyViewBuilder.cpp" />
@@ -214,7 +214,7 @@ copy /y "$(ProjectDir)Resources\ModelExplorerIcon.png" "$(OutDir)ModelExplorerIc
       <Outputs Condition="'$(Configuration)|$(Platform)'=='Release|x64'">.\GeneratedFiles\$(ConfigurationName)\moc_%(Filename).cpp</Outputs>
       <Command Condition="'$(Configuration)|$(Platform)'=='Release|x64'">"$(QTDIR)\bin\moc.exe"  "%(FullPath)" -o ".\GeneratedFiles\$(ConfigurationName)\moc_%(Filename).cpp"  -DUNICODE -DWIN32 -DWIN64 -DQT_DLL -DQT_NO_DEBUG -DNDEBUG -DQT_CORE_LIB -DMODELEXPLORERPLUGIN_LIB -D_WINDLL  "-I.\GeneratedFiles" "-I$(QTDIR)\include" "-I.\GeneratedFiles\$(ConfigurationName)\." "-I.\..\QtPropertyBrowser" "-I." "-I$(renga_com_sdk)\C++\include" "-I$(renga_com_sdk)\tlb" "-fstdafx.h" "-f../../PluginToolButtons.h"</Command>
     </CustomBuild>
-    <ClInclude Include="PropertyManagers.h" />
+    <ClInclude Include="PropertyManager.h" />
     <CustomBuild Include="PropertyView.h">
       <AdditionalInputs Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">$(QTDIR)\bin\moc.exe;%(FullPath)</AdditionalInputs>
       <Message Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Moc%27ing PropertyView.h...</Message>
