+++ ModelExplorerPlugin/PropertyManager.h
